Integrate bus modules into main script
